% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ROCplots.R
\name{ROCplots}
\alias{ROCplots}
\title{ROCplots}
\usage{
ROCplots(data, N = 1000, prev = 0.2, diffInSD = 2,
  whichPlots = c("density", "raw", "ROC", "pv", "nnt"), NNTlower = 3,
  NNTupper = 10)
}
\arguments{
\item{data}{Data frame with columns "class" (binary target variable) and "X" (predictor).}

\item{N}{Sample size}

\item{prev}{Prevalence}

\item{whichPlots}{Which plots to do. Options are c("density", "raw", "ROC", "pv", "nnt")}

\item{NNTlower}{Subjective input. If NNT < NNTlower, the decision is clearly to Treat .}

\item{NNTupper}{Subjective input. If NNT > NNTupper, the decision is clearly to Wait .}

\item{diffInSd}{Difference: E(X | group=1) - E(X | group=0),measured in units of S.D (common to the 2 groups).}
}
\description{
Plots for a binary target and a single continuous predictor.
}

