% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{run}
\usage{
run(shinyDir)
}
\arguments{
\item{shinyDir}{Current options are "shinyElicit" and "shinyContraBayes"}
}
\description{
Run a shiny app for this package.
}
\details{
Currently, runElicit is built from the combinePlots branch. It does not walk
through the design process. The master branch has a walk-through of the steps,
but is not yet supported. When that is activated, the combinePlots version will
be moved to a new directory, shinyCombinePlots.
}

